Running version of random Sauspiel

+++ Game/shuffler.cpp
@@ -34,7 +34,7 @@ std::vector<Card> Shuffler::provideCards() {
 	std::vector<Card> providedCards;
 	for (int i = 0; i < 8; i++) {
 		int nCardsLeft = m_availableCards.size();
-		double idx = rand() * nCardsLeft / RAND_MAX;
+		int idx = (double) rand() * nCardsLeft / RAND_MAX;
 		providedCards.push_back(m_availableCards[idx]);
 		m_availableCards.erase(m_availableCards.begin() + idx);
 	}
